Consumer/requirements.txt
msgpack==1.0.0
kafka-python==2.0.1